\name{birta-package}
\alias{birta-package}
\docType{package}
\title{
Joint Bayesian Inference of miRNA and Transcription Factor Activities
}
\description{
Expression levels of mRNA molecules are regulated by different processes, comprising inhibition or activation by transcription factors and post-transcriptional degradation by microRNAs. birta uses the regulatory networks of TFs and miRNAs together with mRNA and miRNA expression data to predict switches in regulatory activity between two conditions. A Bayesian network is used to model the regulatory structure and Markov-Chain-Monte-Carlo is applied to sample the activity states.
}
\details{
\tabular{ll}{
Package: \tab birta\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2012-03-6\cr
License: \tab GNU GENERAL PUBLIC LICENSE (GPL) >= 2 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Benedikt Zacher, Khalid Abnaof, Stephan Gade, Erfan Younesi, Achim Tresch, Holger Frohlich
Maintainer: Benedikt Zacher <zacher@lmb.uni-muenchen.de>
}
\references{
B. Zacher, K. Abnaof, S. Gade, E. Younesi, A. Tresch, and H. Frohlich. Joint Bayesian Inference of Condition Specific miRNA and Transcription Factor Activities from Combined Gene and microRNA Expression Data. \emph{submitted}, 2012.
}
\keyword{ package }

