\name{potential_swaps}
\docType{data}
\alias{potential_swaps}
\title{Potential swap moves for TF-target and miRNA-target networks (see genestes data set).}
\description{
Potential swap moves for TF-target and miRNA-target networks (see genestes data set) used for the simulation in the vignette.
}
\usage{potential_swaps}
\format{A list containing potential swap moves for the TF- and miRNA-target graph used in the simulation.}
\keyword{datasets}
