\name{sim}
\docType{data}
\alias{sim}
\title{Simulated expression data for mRNAs and miRNAs.}
\description{
  This data set contains simulated expression values of miRNAs and mRNAs, together with the associated TF- and miRNA-target networks. 
}
\usage{sim}
\format{A list  containing the entries dat.mRNA (mRNA expression), dat.miRNA (miRNA expression), miRNAstates (activity states of miRNAs), TFstates (activity states of TFs, TRANSFAC), genesetsmiRNA (miRNA-target graph, targetScan),  genesetsTFs (TF-taget graph).}
\source{Simulation (see reference for details).}
\references{
B. Zacher, K. Abnaof, S. Gade, E. Younesi, A. Tresch, and H. Froehlich. Joint Bayesian Inference of Condition Specific miRNA and Transcription Factor Activities from Combined Gene and microRNA Expression Data. \emph{submitted}, 2012.
}
\keyword{datasets}
